Measures performed

+++ firmware/control.cpp
@@ -6,7 +6,7 @@ static control controlStruct;
 // Returns the spin direction the motor needs to follow in order to respect its limit angles
 int8 choose_direction(motor * pMot);
 
-int8 viable_direction(motor * pMot, int32 pDiffToGoal);
+int8 viable_direction(motor * pMot, int16 pDiffToGoal);
 
 control * get_control_struct() {
     return &controlStruct;
@@ -140,18 +140,18 @@ void control_tick_P_on_torque(motor * pMot) {
 //    controlStruct.deltaAverageCurrent = pMot->targetCurrent - pMot->averageCurrent;
 
 //      // /!\ the -1 conspiracy continues
-//    int32 command = - controlStruct.deltaAverageCurrent * INITIAL_TORQUE_P_COEF;
+//    int16 command = - controlStruct.deltaAverageCurrent * INITIAL_TORQUE_P_COEF;
 
-	int32 command = (-pMot->outputTorque + pMot->targetTorque) * controlStruct.torquePCoef;
+	int16 command = (-pMot->outputTorque + pMot->targetTorque) * controlStruct.torquePCoef;
 	motor_set_command(command);
 }
 
 /**
  * Returns the signed difference between 2 angles
  */
-int32 control_angle_diff(int32 a, int32 b) {
-    int32 diff = a - b;
-    int32 limit = (MAX_ANGLE+1)/2;
+int16 control_angle_diff(int16 a, int16 b) {
+    int16 diff = a - b;
+    int16 limit = (MAX_ANGLE+1)/2;
 
     if (diff > limit) {
         return diff - (MAX_ANGLE + 1);
@@ -166,8 +166,8 @@ int32 control_angle_diff(int32 a, int32 b) {
 /**
  * Returns the other angle between 2 angles (the bigger one, aka the one that is bigger than MAX_ANGLE/2)
  */
-int32 control_other_angle_diff(int32 a, int32 b) {
-    int32 diff = control_angle_diff(a, b);
+int16 control_other_angle_diff(int16 a, int16 b) {
+    int16 diff = control_angle_diff(a, b);
     if (diff > 0) {
     	return diff - (MAX_ANGLE + 1);
     } else {
@@ -193,7 +193,7 @@ int8 choose_direction(motor * pMot) {
         return -1;
     }
 
-    int32 diff = control_angle_diff(pMot->targetAngle, pMot->angle);
+    int16 diff = control_angle_diff(pMot->targetAngle, pMot->angle);
     if (diff == 0) {
         return 0;
     } else {
@@ -202,8 +202,8 @@ int8 choose_direction(motor * pMot) {
 }
 
 
-int8 viable_direction(motor * pMot, int32 pDiffToGoal) {
-    int32 diffToLimit = 0;
+int8 viable_direction(motor * pMot, int16 pDiffToGoal) {
+    int16 diffToLimit = 0;
 
     if (pDiffToGoal > 0) {
             // We are spining positively
